find: First tests
Since the implementation is dependent on these,
I figured I'd send tests first.
Most important are the unicode tests. The program seems to do
a lot of conversion, dependent on the local codepage.
It doesn't seem to detect anything else but utf16le,
for anything else it just treats it like its current codepage -
which is enough to find ANSI strings.

Signed-off-by: Fabian Maurer <dark.shadow4@web.de>