ddraw/tests: Add a test for D3DMCS_COLOR{1/2}.
A ddraw4 version is on my TODO list. ddraw1 and ddraw2 cannot use
lighting and per-vertex color at the same time due to the predefined
D3D*VERTEX structs. ddraw4 can, and it has D3DLIGHTSTATE_COLORVERTEX,
but the material source states are missing. I assume that one or more of
the material properties then track the vertex color, but I have to find
out which.

Signed-off-by: Stefan Dösinger <stefan@codeweavers.com>